Change directory structure and add submodules
