+++ .gitmodules
@@ -1,9 +1,15 @@
+[submodule "libraries/aws-iot-device-sdk-embedded-C"]
+	path = libraries/aws-iot-device-sdk-embedded-C
+	url = https://github.com/aws/aws-iot-device-sdk-embedded-C.git
 [submodule "libraries/ioLibrary_Driver"]
 	path = libraries/ioLibrary_Driver
 	url = https://github.com/Wiznet/ioLibrary_Driver.git
 [submodule "libraries/mbedtls"]
 	path = libraries/mbedtls
 	url = https://github.com/ARMmbed/mbedtls.git
-[submodule "libraries/aws-iot-device-sdk-embedded-C"]
-	path = libraries/aws-iot-device-sdk-embedded-C
-	url = https://github.com/aws/aws-iot-device-sdk-embedded-C.git
+[submodule "libraries/pico-extras"]
+	path = libraries/pico-extras
+	url = https://github.com/raspberrypi/pico-extras.git
+[submodule "libraries/pico-sdk"]
+	path = libraries/pico-sdk
+	url = https://github.com/raspberrypi/pico-sdk.git
