+++ CMakeLists.txt
@@ -2,8 +2,14 @@
 cmake_minimum_required(VERSION 3.12)
 
 # Pull in SDK (must be before project)
-include(pico_sdk_import.cmake)
+if(NOT DEFINED PICO_SDK_PATH)
+    set(PICO_SDK_PATH ${CMAKE_SOURCE_DIR}/libraries/pico-sdk)
+    message(STATUS "PICO_SDK_PATH = ${PICO_SDK_PATH}")
+endif()
+
 include(rp2040_hat_aws_c-patch.cmake)
+include(pico_sdk_import.cmake)
+include(rp2040_hat_aws_c_sdk_version.cmake)
 
 # Set project name
 set(PROJECT_NAME RP2040-HAT-AWS-C)
@@ -17,22 +23,46 @@ set(CMAKE_CXX_STANDARD 17)
 # Initialize the SDK
 pico_sdk_init()
 
-add_definitions(-DMBEDTLS_CONFIG_FILE="${CMAKE_SOURCE_DIR}/interface/ssl_config.h")
-add_definitions(-DHTTP_DO_NOT_USE_CUSTOM_CONFIG)
-add_definitions(-DMQTT_DO_NOT_USE_CUSTOM_CONFIG)
+# Set the project root directory if it's not already defined, as may happen if
+# the tests folder is included directly by a parent project, without including
+# the top level CMakeLists.txt.
+if(NOT DEFINED AWS_SDK_DIR)
+    set(AWS_SDK_DIR ${CMAKE_SOURCE_DIR}/libraries/aws-iot-device-sdk-embedded-C)
+    message(STATUS "AWS_SDK_DIR = ${AWS_SDK_DIR}")
+endif()
+
+if(NOT DEFINED WIZNET_DIR)
+    set(WIZNET_DIR ${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver)
+    message(STATUS "WIZNET_DIR = ${WIZNET_DIR}")
+endif()
+
+if(NOT DEFINED MBEDTLS_LIB_DIR)
+    set(MBEDTLS_LIB_DIR ${CMAKE_SOURCE_DIR}/libraries/mbedtls)
+    message(STATUS "MBEDTLS_LIB_DIR = ${MBEDTLS_LIB_DIR}")
+endif()
+
+if(NOT DEFINED PORT_DIR)
+    set(PORT_DIR ${CMAKE_SOURCE_DIR}/port)
+    message(STATUS "PORT_DIR = ${PORT_DIR}")
+endif()
+
+# Turn off mbedtls test mode 
+set(ENABLE_PROGRAMS OFF CACHE BOOL "Build mbedtls programs")
+set(ENABLE_TESTING OFF CACHE BOOL "Build mbedtls testing")
+add_definitions(-DMBEDTLS_CONFIG_FILE="${PORT_DIR}/mbedtls/inc/ssl_config.h")
+add_definitions(-DSET_TRUSTED_CERT_IN_SAMPLES)
 
 # Hardware-specific examples in subdirectories:
 add_subdirectory(examples)
 
-# Add interface in subdirectories
-add_subdirectory(interface)
-
 # Add libraries in subdirectories
-add_subdirectory(libraries)
+add_subdirectory(${CMAKE_SOURCE_DIR}/libraries)
+add_subdirectory(${MBEDTLS_LIB_DIR})
+add_subdirectory(${PORT_DIR})
 
 # Set compile options
 add_compile_options(
-		-Wall
+        -Wall
         -Wno-format          # int != int32_t as far as the compiler is concerned because gcc has int32_t as long int
         -Wno-unused-function # we have some for the docs that aren't called
         -Wno-maybe-uninitialized
