--- Interface/dns_config.h
@@ -1,14 +1,0 @@
-/**
- * Copyright (c) 2021 WIZnet Co.,Ltd
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#ifndef _DNS_CONFIG_H_
-#define _DNS_CONFIG_H_
-
-#define DNS_SOCKET_NUM 3
-
-uint8_t dns_server_ip[4] = {8, 8, 8, 8};
-
-#endif /* _DNS_CONFIG_H_ */
