+++ examples/aws_iot_http/http_certificate.h
