+++ examples/aws_iot_mqtt/mqtt_certificate.h
