+++ patches/02_aws_iot_device_sdk_embedded_c_corehttp_network_interface.patch
