--- patches/02_iolibrary_driver_ftp_client.patch
@@ -1,31 +1,0 @@
-diff --git a/Internet/FTPClient/ftpc.h b/Internet/FTPClient/ftpc.h
-index ef2be52..93c6f6f 100644
---- a/Internet/FTPClient/ftpc.h
-+++ b/Internet/FTPClient/ftpc.h
-@@ -22,10 +22,9 @@ extern "C" {
- /* If your target support a file system, you have to activate this feature and implement. */
- //#define F_FILESYSTEM
- 
--/* Change to your Chipset Uart function, you have to activate this feature and implement.
-- * Change!! -> Board_UARTGetCharBlocking()
-- * Below is an example of a function of lpc_chip library. */
--//#define ftp_getc()	Board_UARTGetCharBlocking()
-+/* Change to your chipset UART function, you have to activate this feature and implement.
-+ * Change macro to suit your development environment. */
-+#define ftp_getc()	getchar()
- 
- #ifdef F_FILESYSTEM
- #include "ff.h"
-@@ -34,10 +33,8 @@ extern "C" {
- #ifndef	ftp_getc()
- #define Need_UARTGetCharBlocking_func
- #else
--/* Change library
-- * Change!! -> board_api.h,
-- * Below is an example of a function of lpc_chip library. */
--#include "board_api.h"
-+/* Change library to suit your development environment. */
-+//#include "board_api.h"
- #endif
- 
-
