+++ port/aws-iot-device-sdk-embedded-C/inc/core_http_config.h
