+++ port/aws-iot-device-sdk-embedded-C/inc/http_transport_interface.h
@@ -34,12 +34,6 @@
  * Variables
  * ----------------------------------------------------------------------------------------------------
  */
-
-/*
- * ----------------------------------------------------------------------------------------------------
- * Struct
- * ----------------------------------------------------------------------------------------------------
- */
 typedef enum
 {
     HTTP_IDLE = 0,
