+++ port/aws-iot-device-sdk-embedded-C/src/dns_interface.c
@@ -4,14 +4,31 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Includes
+  * ----------------------------------------------------------------------------------------------------
+  */
 #include "socket.h"
-#include "timer_interface.h"
 
 #include "dns.h"
 #include "dns_interface.h"
 #include "dns_config.h"
+
+#include "timer_interface.h"
 #include "util.h"
 
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Macros
+  * ----------------------------------------------------------------------------------------------------
+  */
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Variables
+  * ----------------------------------------------------------------------------------------------------
+  */
 /* Header for all domain messages */
 struct dhdr_handler
 {
@@ -48,6 +65,11 @@ extern uint16_t DNS_MSGID;   // DNS message ID
 extern uint32_t dns_1s_tick; // for timout of DNS processing
 extern uint8_t g_common_buf[];
 
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Functions
+  * ----------------------------------------------------------------------------------------------------
+  */
 /* DNS CLIENT INIT */
 void DNS_init_handler(uint8_t s, uint8_t *buf, uint32_t (*tickFunc)(void))
 {
