+++ port/aws-iot-device-sdk-embedded-C/src/http_transport_interface.c
@@ -17,9 +17,16 @@
 #include "http_transport_interface.h"
 #include "http_parser.h"
 #include "core_http_client_private.h"
-#include "timer_interface.h"
+
 #include "dns_interface.h"
+#include "timer_interface.h"
 #include "util.h"
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Macros
+  * ----------------------------------------------------------------------------------------------------
+  */
 
 /*
  * ----------------------------------------------------------------------------------------------------
