+++ port/aws-iot-device-sdk-embedded-C/src/ssl_transport_interface.c
@@ -4,8 +4,14 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+/*
+ * ----------------------------------------------------------------------------------------------------
+ * Includes
+ * ----------------------------------------------------------------------------------------------------
+ */
 #include <stdio.h>
 #include <string.h>
+#include <stdbool.h>
 
 #include "mbedtls/x509_crt.h"
 #include "mbedtls/error.h"
@@ -13,10 +19,29 @@
 #include "mbedtls/ctr_drbg.h"
 #include "mbedtls/compat-1.3.h"
 
-#include "ssl_transport_interface.h"
-#include "timer_interface.h"
 #include "socket.h"
 
+#include "ssl_transport_interface.h"
+
+#include "timer_interface.h"
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Macros
+  * ----------------------------------------------------------------------------------------------------
+  */
+
+/*
+ * ----------------------------------------------------------------------------------------------------
+ * Variables
+ * ----------------------------------------------------------------------------------------------------
+ */
+
+/*
+ * ----------------------------------------------------------------------------------------------------
+ * Functions
+ * ----------------------------------------------------------------------------------------------------
+ */
 int ssl_random_callback(void *p_rng, unsigned char *output, size_t output_len)
 {
     int i;
