+++ port/aws-iot-device-sdk-embedded-C/src/timer_interface.c
@@ -1,0 +1,53 @@
+/**
+ * Copyright (c) 2021 WIZnet Co.,Ltd
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+/*
+ * ----------------------------------------------------------------------------------------------------
+ * Includes
+ * ----------------------------------------------------------------------------------------------------
+ */
+#include "timer_interface.h"
+
+#include "dhcp.h"
+#include "dns.h"
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Macros
+  * ----------------------------------------------------------------------------------------------------
+  */
+
+/*
+ * ----------------------------------------------------------------------------------------------------
+ * Variables
+ * ----------------------------------------------------------------------------------------------------
+ */
+volatile uint32_t devtime_msec = 0;
+volatile uint32_t g_msec_cnt = 0;
+
+/*
+ * ----------------------------------------------------------------------------------------------------
+ * Functions
+ * ----------------------------------------------------------------------------------------------------
+ */
+void repeating_timer_callback(void)
+{
+    devtime_msec++;
+    g_msec_cnt++;
+
+    if (g_msec_cnt >= 1000 - 1)
+    {
+        g_msec_cnt = 0;
+
+        DHCP_time_handler();
+        DNS_time_handler();
+    }
+}
+
+time_t millis(void)
+{
+    return devtime_msec;
+}
