+++ port/ioLibrary_Driver/w5x00_spi.h
@@ -1,0 +1,190 @@
+/**
+ * Copyright (c) 2021 WIZnet Co.,Ltd
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#ifndef _W5X00_SPI_H_
+#define _W5X00_SPI_H_
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Macros
+  * ----------------------------------------------------------------------------------------------------
+  */
+/* SPI */
+#define SPI_PORT spi0
+
+#define PIN_SCK 18
+#define PIN_MOSI 19
+#define PIN_MISO 16
+#define PIN_CS 17
+#define PIN_RST 20
+
+/* Use SPI DMA */
+//#define USE_SPI_DMA // if you want to use SPI DMA, uncomment.
+
+/* Clock */
+#define PLL_SYS_KHZ (133 * 1000)
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Functions
+  * ----------------------------------------------------------------------------------------------------
+  */
+/* W5x00 */
+/*! \brief Set CS pin
+ *  \ingroup w5x00_spi
+ * 
+ * Set chip select pin of spi0 to low(Active low).
+ * 
+ * \param none
+ */
+static inline void wizchip_select(void);
+
+/*! \brief Set CS pin
+ *  \ingroup w5x00_spi
+ * 
+ * Set chip select pin of spi0 to high(Inactive high).
+ * 
+ * \param none
+ */
+static inline void wizchip_deselect(void);
+
+/*! \brief Read from an SPI device, blocking
+ *  \ingroup w5x00_spi
+ * Set spi_read_blocking function.
+ * Read byte from SPI to rx_data buffer.
+ * Blocks until all data is transferred. No timeout, as SPI hardware always transfers at a known data rate.
+ * 
+ * \param none
+ */
+static uint8_t wizchip_read(void);
+
+/*! \brief Write to an SPI device, blocking
+ *  \ingroup w5x00_spi
+ * Set spi_write_blocking function.
+ * Write byte from tx_data buffer to SPI device. 
+ * Blocks until all data is transferred. No timeout, as SPI hardware always transfers at a known data rate.
+ * 
+ * \param tx_data Buffer of data to write
+ */
+static void wizchip_write(uint8_t tx_data);
+#ifdef USE_SPI_DMA
+/*! \brief Configure all DMA parameters and optionally start transfer
+ *  \ingroup w5x00_spi
+ * 
+ * Configure all DMA parameters and read from DMA  
+ * 
+ * \param pBuf Buffer of data to read
+ * \param len element count (each element is of size transfer_data_size)
+ */
+static void wizchip_read_burst(uint8_t *pBuf, uint16_t len);
+
+/*! \brief Configure all DMA parameters and optionally start transfer
+ *  \ingroup w5x00_spi
+ * 
+ * Configure all DMA parameters and write to DMA
+ * 
+ * \param pBuf Buffer of data to write
+ * \param len element count (each element is of size transfer_data_size)
+ */
+static void wizchip_write_burst(uint8_t *pBuf, uint16_t len);
+#endif
+
+/*! \brief Enter a critical section
+ *  \ingroup w5x00_spi
+ * Set ciritical section enter blocking function.
+ * If the spin lock associated with this critical section is in use, then this
+ * method will block until it is released.
+ * 
+ * \param none
+ */
+static void wizchip_critical_section_lock(void);
+
+/*! \brief Release a critical section
+ *  \ingroup w5x00_spi
+ * Set ciritical section exit function.
+ * Release a critical section.
+ * 
+ * \param none
+ */
+static void wizchip_critical_section_unlock(void);
+
+/*! \brief Initialize SPI instances and Set DMA channel
+ *  \ingroup w5x00_spi
+ * Set GPIO to spi0.
+ * Puts the SPI into a known state, and enable it.
+ * Set DMA channel completion channel.
+ * 
+ * \param none
+ */
+void wizchip_spi_initialize(void);
+
+/*! \brief Initialize a critical section structure
+ *  \ingroup w5x00_spi
+ * The critical section is initialized ready for use.
+ * Registers callback function for critical section for WIZchip.
+ * 
+ * \param none
+ */
+void wizchip_cris_initialize(void);
+
+/*! \brief W5x00 chip reset
+ *  \ingroup w5x00_spi
+ * 
+ * Set a reset pin and reset.
+ *
+ * \param none
+ */
+void wizchip_reset(void);
+
+/*! \brief Initialize WIZchip
+ *  \ingroup w5x00_spi
+ * Set callback function to read/write byte using SPI.
+ * Set callback function for WIZchip select/deselect.
+ * Set memory size of W5x00 chip and monitor PHY link status.
+ * 
+ * \param none
+ */
+void wizchip_initialize(void);
+
+/*! \brief Check chip version
+ *  \ingroup w5x00_spi
+ * 
+ * Get version information.
+ * 
+ * \param none
+ */
+void wizchip_check(void);
+
+/* Network */
+/*! \brief Initialize network
+ *  \ingroup w5x00_spi
+ * 
+ * Set network information.
+ * 
+ * \param net_info network information.
+ */
+void network_initialize(wiz_NetInfo net_info);
+
+/*! \brief Print network information
+ *  \ingroup w5x00_spi
+ * 
+ * Print network information about MAC address, IP address, Subnet mask, Gateway, DHCP and DNS address.
+ * 
+ * \param net_info network information.
+ */
+void print_network_information(wiz_NetInfo net_info);
+
+/* Clock */
+/*! \brief Attempt to set a system clock frequency in khz
+ *  \ingroup w5x00_spi
+ * Set a system clock frequency in khz.
+ * Configure the specified clock.
+ * 
+ * \param none
+ */
+void set_clock_khz(void);
+
+#endif /* _W5X00_SPI_H_ */
