+++ port/port_common.h
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2021 WIZnet Co.,Ltd
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#ifndef _PORT_COMMON_H_
+#define _PORT_COMMON_H_
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Includes
+  * ----------------------------------------------------------------------------------------------------
+  */
+/* Common */
+#include "pico/stdlib.h"
+#include "pico/binary_info.h"
+#include "pico/critical_section.h"
+#include "hardware/spi.h"
+#include "hardware/dma.h"
+#include "hardware/clocks.h"
+
+#endif /* _PORT_COMMON_H_ */
