+++ port/timer/timer.h
@@ -1,0 +1,52 @@
+/**
+ * Copyright (c) 2021 WIZnet Co.,Ltd
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#ifndef _TIMER_H_
+#define _TIMER_H_
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Macros
+  * ----------------------------------------------------------------------------------------------------
+  */
+/* Timeout */
+#define RECV_TIMEOUT (1000 * 10) // 10 seconds
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Functions
+  * ----------------------------------------------------------------------------------------------------
+  */
+/* Timer */
+/*! \brief Initialize timer callback function
+ * \ingroup timer
+ *
+ * Add a repeating timer that is called repeatedly at the specified interval in microseconds.
+ *
+ * \param callback the repeating timer callback function
+ */
+void wizchip_1ms_timer_initialize(void (*callback)(void));
+
+/*! \brief Assign timer callback function
+ * \ingroup timer
+ *
+ * 1ms timer callback function.
+ * 
+ * \param t Information about a repeating timer
+ */
+bool wizchip_1ms_timer_callback(struct repeating_timer *t);
+
+/* Delay */
+/*! \brief Wait for the given number of milliseconds before returning
+ * \ingroup timer
+ *
+ * This method attempts to perform a lower power sleep (using WFE) as much as possible.
+ *
+ * \param ms the number of milliseconds to sleep
+ */
+void wizchip_delay_ms(uint32_t ms);
+
+#endif /* _TIMER_H_ */
