+++ rp2040_hat_aws_c-patch.cmake
@@ -11,28 +11,18 @@ endif()
 message("RP2040-HAT-AWS-C patch utils found")
 
 set(RP2040_HAT_AWS_C_SRC_DIR "${CMAKE_CURRENT_SOURCE_DIR}")
-set(IOLIBRARY_DRIVER_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/ioLibrary_Driver")
-set(MBEDTLS_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/mbedtls")
 set(AWS_IOT_DEVICE_SDK_EMBEDDED_C_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/aws-iot-device-sdk-embedded-C")
 set(AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/aws-iot-device-sdk-embedded-C/libraries/standard/coreHTTP")
 set(AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREMQTT_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/aws-iot-device-sdk-embedded-C/libraries/standard/coreMQTT")
-set(AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_HTTP_PARSER_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/aws-iot-device-sdk-embedded-C/libraries/standard/coreHTTP/dependency/3rdparty/http_parser")
+set(AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_HTTP_PARSER_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/aws-iot-device-sdk-embedded-C/libraries/standard/coreHTTP/source/dependency/3rdparty/http_parser")
+set(IOLIBRARY_DRIVER_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/ioLibrary_Driver")
+set(MBEDTLS_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/mbedtls")
+set(PICO_EXTRAS_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/pico-extras")
+set(PICO_SDK_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/pico-sdk")
+set(PICO_SDK_TINYUSB_SRC_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/libraries/pico-sdk/lib/tinyusb")
 set(RP2040_HAT_AWS_C_PATCH_DIR "${RP2040_HAT_AWS_C_SRC_DIR}/patches")
 
-if(EXISTS "${IOLIBRARY_DRIVER_SRC_DIR}/.git")
-	message("cleaning ioLibrary_Driver...")
-	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${IOLIBRARY_DRIVER_SRC_DIR} clean -fdx)
-	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${IOLIBRARY_DRIVER_SRC_DIR} reset --hard)
-	message("ioLibrary_Driver cleaned")
-endif()
-
-if(EXISTS "${MBEDTLS_SRC_DIR}/.git")
-	message("cleaning mbedtls...")
-	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${MBEDTLS_SRC_DIR} clean -fdx)
-	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${MBEDTLS_SRC_DIR} reset --hard)
-	message("mbedtls cleaned")
-endif()
-
+# Delete untracked files in aws-iot-device-sdk-embedded-C
 if(EXISTS "${AWS_IOT_DEVICE_SDK_EMBEDDED_C_SRC_DIR}/.git")
 	message("cleaning aws-iot-device-sdk-embedded-C...")
 	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${AWS_IOT_DEVICE_SDK_EMBEDDED_C_SRC_DIR} clean -fdx)
@@ -40,8 +30,41 @@ if(EXISTS "${AWS_IOT_DEVICE_SDK_EMBEDDED_C_SRC_DIR}/.git")
 	message("aws-iot-device-sdk-embedded-C cleaned")
 endif()
 
+# Delete untracked files in ioLibrary_Driver
+if(EXISTS "${IOLIBRARY_DRIVER_SRC_DIR}/.git")
+	message("cleaning ioLibrary_Driver...")
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${IOLIBRARY_DRIVER_SRC_DIR} clean -fdx)
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${IOLIBRARY_DRIVER_SRC_DIR} reset --hard)
+	message("ioLibrary_Driver cleaned")
+endif()
+
+# Delete untracked files in mbedtls
+if(EXISTS "${MBEDTLS_SRC_DIR}/.git")
+	message("cleaning mbedtls...")
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${MBEDTLS_SRC_DIR} clean -fdx)
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${MBEDTLS_SRC_DIR} reset --hard)
+	message("mbedtls cleaned")
+endif()
+
+# Delete untracked files in pico-extras
+if(EXISTS "${PICO_EXTRAS_SRC_DIR}/.git")
+	message("cleaning pico-extras...")
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${PICO_EXTRAS_SRC_DIR} clean -fdx)
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${PICO_EXTRAS_SRC_DIR} reset --hard)
+	message("pico-extras cleaned")
+endif()
+
+# Delete untracked files in pico-sdk
+if(EXISTS "${PICO_SDK_SRC_DIR}/.git")
+	message("cleaning pico-sdk...")
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${PICO_SDK_SRC_DIR} clean -fdx)
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${PICO_SDK_SRC_DIR} reset --hard)
+	message("pico-sdk cleaned")
+endif()
+
 execute_process(COMMAND ${GIT_EXECUTABLE} -C ${RP2040_HAT_AWS_C_SRC_DIR} submodule update --init)
 
+# Delete untracked files in coreHTTP
 if(EXISTS "${AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_SRC_DIR}/.git")
 	message("cleaning aws-iot-device-sdk-embedded-C coreHTTP...")
 	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_SRC_DIR} clean -fdx)
@@ -49,6 +72,7 @@ if(EXISTS "${AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_SRC_DIR}/.git")
 	message("aws-iot-device-sdk-embedded-C coreHTTP cleaned")
 endif()
 
+# Delete untracked files in coreMQTT
 if(EXISTS "${AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREMQTT_SRC_DIR}/.git")
 	message("cleaning aws-iot-device-sdk-embedded-C coreMQTT...")
 	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREMQTT_SRC_DIR} clean -fdx)
@@ -58,6 +82,7 @@ endif()
 
 execute_process(COMMAND ${GIT_EXECUTABLE} -C ${AWS_IOT_DEVICE_SDK_EMBEDDED_C_SRC_DIR} submodule update --init)
 
+# Delete untracked files in http_parser
 if(EXISTS "${AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_HTTP_PARSER_SRC_DIR}/.git")
 	message("cleaning aws-iot-device-sdk-embedded-C coreHTTP http_parser...")
 	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_HTTP_PARSER_SRC_DIR} clean -fdx)
@@ -67,12 +92,21 @@ endif()
 
 execute_process(COMMAND ${GIT_EXECUTABLE} -C ${AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_SRC_DIR} submodule update --init)
 
+# Delete untracked files in tinyusb
+if(EXISTS "${PICO_SDK_TINYUSB_SRC_DIR}/.git")
+	message("cleaning pico-sdk tinyusb...")
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${PICO_SDK_TINYUSB_SRC_DIR} clean -fdx)
+	execute_process(COMMAND ${GIT_EXECUTABLE} -C ${PICO_SDK_TINYUSB_SRC_DIR} reset --hard)
+	message("pico-sdk tinyusb cleaned")
+endif()
+
+execute_process(COMMAND ${GIT_EXECUTABLE} -C ${PICO_SDK_SRC_DIR} submodule update --init)
+
 # ioLibrary_Driver patch
 message("submodules ioLibrary_Driver initialised")
 
 file(GLOB IOLIBRARY_DRIVER_PATCHES 
 	"${RP2040_HAT_AWS_C_PATCH_DIR}/01_iolibrary_driver_ethernet_chip.patch" 
-	"${RP2040_HAT_AWS_C_PATCH_DIR}/02_iolibrary_driver_ftp_client.patch"
 	)
 
 foreach(IOLIBRARY_DRIVER_PATCH IN LISTS IOLIBRARY_DRIVER_PATCHES)
@@ -83,26 +117,11 @@ foreach(IOLIBRARY_DRIVER_PATCH IN LISTS IOLIBRARY_DRIVER_PATCHES)
 	)
 endforeach()
 
-# mbedtls patch
-message("submodules mbedtls initialised")
-
-file(GLOB MBEDTLS_PATCHES 
-	"${RP2040_HAT_AWS_C_PATCH_DIR}/03_mbedtls_test_mode.patch"
-	)
-
-foreach(MBEDTLS_PATCH IN LISTS MBEDTLS_PATCHES)
-	message("Running patch ${MBEDTLS_PATCH}")
-	execute_process(
-		COMMAND ${GIT_EXECUTABLE} apply --ignore-whitespace ${MBEDTLS_PATCH}
-		WORKING_DIRECTORY ${MBEDTLS_SRC_DIR}
-	)
-endforeach()
-
-# aws-iot-device-sdk-embedded-C coreHTTP patch
-message("submodules aws-iot-device-sdk-embedded-C initialised")
+# coreHTTP patch
+message("submodules aws-iot-device-sdk-embedded-C coreHTTP initialised")
 
 file(GLOB AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_PATCHES 
-	"${RP2040_HAT_AWS_C_PATCH_DIR}/04_aws_iot_device_sdk_embedded_c_corehttp_network_interface.patch"
+	"${RP2040_HAT_AWS_C_PATCH_DIR}/02_aws_iot_device_sdk_embedded_c_corehttp_network_interface.patch"
 	)
 
 foreach(AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_PATCH IN LISTS AWS_IOT_DEVICE_SDK_EMBEDDED_C_COREHTTP_PATCHES)
