+++ rp2040_hat_aws_c_sdk_version.cmake
@@ -1,0 +1,10 @@
+# RP2040-HAT-AWS-C_BUILD_DEFINE: RP2040-HAT-AWS-C_SDK_VERSION_MAJOR, SDK major version number
+set(RP2040-HAT-AWS-C_SDK_VERSION_MAJOR 2)
+# RP2040-HAT-AWS-C_BUILD_DEFINE: RP2040-HAT-AWS-C_SDK_VERSION_MINOR, SDK minor version number
+set(RP2040-HAT-AWS-C_SDK_VERSION_MINOR 0)
+# RP2040-HAT-AWS-C_BUILD_DEFINE: RP2040-HAT-AWS-C_SDK_VERSION_REVISION, SDK version revision
+set(RP2040-HAT-AWS-C_SDK_VERSION_REVISION 0)
+# RP2040-HAT-AWS-C_BUILD_DEFINE: RP2040-HAT-AWS-C_SDK_VERSION_STRING, SDK version
+set(RP2040-HAT-AWS-C_SDK_VERSION_STRING "${RP2040-HAT-AWS-C_SDK_VERSION_MAJOR}.${RP2040-HAT-AWS-C_SDK_VERSION_MINOR}.${RP2040-HAT-AWS-C_SDK_VERSION_REVISION}")
+
+message(STATUS "RP2040-HAT-AWS-C SDK version is ${RP2040-HAT-AWS-C_SDK_VERSION_STRING}")
